Move type files into type folder to make project structure simpler
