+++ frontend-service/src/components/types/MapBoxMap.types.ts
