+++ frontend-service/src/components/types/TopBarGame.types.ts
@@ -1,0 +1,11 @@
+interface TopBarGameProps {
+	roundLocation: string
+	score: number
+	currentRound: number
+	roundCompleted: boolean
+	moveToNextRound: () => void,
+	setIsEndModalOpen: (value: boolean) => void,
+	isEndModalOpen: boolean
+}
+
+export type { TopBarGameProps };
