+++ frontend-service/src/tests/utils/mapUtils.test.ts
@@ -1,6 +1,6 @@
 import { test, expect } from 'vitest';
 import { calculateKm } from '../../utils/mapUtils';
-import type { Pin } from '../../components/Game.types';
+import type { Pin } from '../../components/types/Game.types';
 
 test('Same coordinates distance is 0km', () => {
 	const northPoleCoords: Pin = [0, 90]
